+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - a warning shown when an object cannot be converted into another one
   - a warning shown when an object does not handle a message (for strong-typed ones)
   - #511 : Infinite extension **populate** call
+  - #514 : Incorrect warning message for unhandled function message 
 - TOOLS
    - #323 : Run ELENA code without build